+++ example/lib/game_tiled_map.dart
@@ -17,8 +17,7 @@ class GameTiledMap extends StatelessWidget {
   Widget build(BuildContext context) {
     return LayoutBuilder(
       builder: (context, constraints) {
-        DungeonMap.tileSize = max(constraints.maxHeight, constraints.maxWidth) /
-            (kIsWeb ? 25 : 22);
+        DungeonMap.tileSize = max(constraints.maxHeight, constraints.maxWidth) / (kIsWeb ? 25 : 22);
         return BonfireTiledWidget(
           joystick: Joystick(
             keyboardEnable: true,
@@ -54,20 +53,14 @@ class GameTiledMap extends StatelessWidget {
             'tiled/mapa1.json',
             forceTileSize: Size(DungeonMap.tileSize, DungeonMap.tileSize),
           )
-            ..registerObject(
-                'goblin', (x, y, width, height) => Goblin(Position(x, y)))
-            ..registerObject(
-                'torch', (x, y, width, height) => Torch(Position(x, y)))
-            ..registerObject('barrel',
-                (x, y, width, height) => BarrelDraggable(Position(x, y)))
-            ..registerObject(
-                'spike', (x, y, width, height) => Spikes(Position(x, y)))
-            ..registerObject(
-                'chest', (x, y, width, height) => Chest(Position(x, y))),
+            ..registerObject('goblin', (x, y, width, height) => Goblin(Position(x, y)))
+            ..registerObject('torch', (x, y, width, height) => Torch(Position(x, y)))
+            ..registerObject('barrel', (x, y, width, height) => BarrelDraggable(Position(x, y)))
+            ..registerObject('spike', (x, y, width, height) => Spikes(Position(x, y)))
+            ..registerObject('chest', (x, y, width, height) => Chest(Position(x, y))),
           background: BackgroundColorGame(Colors.blueGrey[900]),
           lightingColorGame: Colors.black.withOpacity(0.5),
-          cameraZoom:
-              1.0, // you can change the game zoom here or directly on camera
+          cameraZoom: 1.0, // you can change the game zoom here or directly on camera
         );
       },
     );