+++ lib/util/camera/camera.dart
@@ -2,6 +2,7 @@ import 'package:bonfire/base/game_component.dart';
 import 'package:bonfire/base/rpg_game.dart';
 import 'package:bonfire/bonfire.dart';
 import 'package:flame/components/mixins/has_game_ref.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 class Camera with HasGameRef<RPGGame> {
@@ -21,8 +22,14 @@ class Camera with HasGameRef<RPGGame> {
 
   Rect get cameraRect => Rect.fromCenter(
         center: Offset(position.x, position.y),
-        width: gameRef.size.width / zoom + 80,
-        height: gameRef.size.height / zoom + 80,
+        width: gameRef.size.width,
+        height: gameRef.size.height,
+      );
+
+  Rect get cameraRectWithMargin => Rect.fromCenter(
+        center: Offset(position.x, position.y),
+        width: gameRef.size.width + 20,
+        height: gameRef.size.height + 20,
       );
 
   void moveTop(double displacement) {
@@ -169,14 +176,10 @@ class Camera with HasGameRef<RPGGame> {
     final verticalDistance = screenCenter.dy - positionTarget.dy;
 
     if (horizontalDistance.abs() > horizontal) {
-      this.position.x += horizontalDistance > 0
-          ? horizontal - horizontalDistance
-          : -horizontalDistance - horizontal;
+      this.position.x += horizontalDistance > 0 ? horizontal - horizontalDistance : -horizontalDistance - horizontal;
     }
     if (verticalDistance.abs() > vertical) {
-      this.position.y += verticalDistance > 0
-          ? vertical - verticalDistance
-          : -verticalDistance - vertical;
+      this.position.y += verticalDistance > 0 ? vertical - verticalDistance : -verticalDistance - vertical;
     }
 
     if (moveOnlyMapArea) {
@@ -209,20 +212,28 @@ class Camera with HasGameRef<RPGGame> {
   bool isComponentOnCamera(GameComponent c) {
     if (gameRef?.size == null || c.position == null) return false;
 
-    return cameraRect.overlaps(c.position);
+    return cameraRectWithMargin.overlaps(c.position);
+  }
+
+  bool isRectOnCamera(Rect c) {
+    if (gameRef?.size == null || c == null) return false;
+
+    return cameraRectWithMargin.overlaps(c);
   }
 
   Offset worldPositionToScreen(Offset position) {
     return position.translate(
-      -this.position.x + gameRef.size.width / 2,
-      -this.position.y + gameRef.size.height / 2,
+      this.cameraRect.left * -1,
+      this.cameraRect.top * -1,
     );
   }
 
-  Offset cameraPositionToWorld(Offset position) {
-    return position.translate(
-      this.position.x - gameRef.size.width / 2,
-      this.position.y - gameRef.size.height / 2,
+  Offset screenPositionToWorld(Offset position) {
+    double diffX = position.dx - gameRef.size.width / 2;
+    double diffY = position.dy - gameRef.size.height / 2;
+    return Offset(
+      this.cameraRect.center.dx + (diffX / zoom),
+      this.cameraRect.center.dy + (diffY / zoom),
     );
   }